fix bug with processing initial dependency artifacts
